requirements.txt
annotated-types==0.7.0
apt-clone==0.2.1
apturl==0.5.2
beautifulsoup4==4.8.2
blinker==1.8.2
blis==0.7.11
Brlapi==0.7.0
catalogue==2.0.10
ccsm==0.9.14.1
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
click-plugins==1.1.1
cloudpathlib==0.18.1
command-not-found==0.3
compizconfig-python==0.9.14.1
confection==0.1.5
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cymem==2.0.8
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
docopt==0.6.2
en-core-web-sm==3.7.1
entrypoints==0.3
filelock==3.15.4
flask==3.0.3
Flask-Cors==4.0.1
flask-sqlalchemy==3.1.1
fsspec==2024.6.1
fugashi==1.3.2
greenlet==3.0.3
grpcio==1.16.1
httplib2==0.14.0
huggingface-hub==0.23.4
idna==2.8
ifaddr==0.1.6
importlib-metadata==8.0.0
ipadic==1.0.0
itsdangerous==2.2.0
ja-core-news-sm==3.7.0
jinja2==3.1.4
joblib==1.4.2
keyring==18.0.1
langcodes==3.4.0
langdetect==1.0.9
language-data==1.2.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.0
Mako==1.1.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
netifaces==0.10.4
nl-core-news-sm==3.7.0
nose==1.3.7
numpy==1.24.4
oauthlib==3.1.0
onboard==1.4.1
openstep-parser==1.5.4
packaging==24.1
pbxproj==4.2.0
pefile==2023.2.7
pexpect==4.6.0
Pillow==7.0.0
preshed==3.0.9
protobuf==3.6.1
psutil==5.5.1
pybind11==2.13.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
pydantic==2.7.4
pydantic-core==2.18.4
pygments==2.18.0
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyNaCl==1.3.0
pyparted==3.11.2
python-apt==2.0.0
python-debian===0.1.36ubuntu1
python-docx==1.1.2
python-magic==0.4.27
python-xapp==2.0.1
python-xlib==0.23
pyxdg==0.26
PyYAML==5.3.1
regex==2024.5.15
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rich==13.7.1
sacremoses==0.1.1
safetensors==0.4.3
SecretStorage==2.3.1
sentencepiece==0.2.0
setproctitle==1.1.10
shellingham==1.5.4
simplejson==3.16.0
six==1.14.0
smart-open==7.0.4
soupsieve==1.9.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
ssh-import-id==5.10
SudachiDict-core==20240409
SudachiPy==0.6.8
systemd-python==234
termcolor==2.4.0
thinc==8.2.5
tika==2.6.0
tokenizers==0.19.1
torch==1.13.1+cpu
torchvision==0.14.1+cpu
tqdm==4.66.4
transformers==4.42.3
typer==0.12.3
typing-extensions==4.12.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
wadllib==1.3.3
wasabi==1.1.3
weasel==0.4.1
werkzeug==3.0.3
wrapt==1.16.0
xkit==0.0.0
zeroconf==0.24.4
zipp==3.19.2
openai>=1.0.0
anthropic>=0.18.0
